# qr-code-checkin-app/api/doc/openapi.yaml
openapi: 3.0.0
info:
  title: QR Code Check-in System API
  version: 1.0.0
  description: API for managing QR code-based event check-ins

servers:
  - url: https://api.example.com/v1

paths:
  /accounts/{accountId}/events:
    post:
      summary: Create a new event
      operationId: createEvent
      tags:
        - events
      parameters:
        - name: accountId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventCreation'
      responses:
        '201':
          description: Event created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '400':
          description: Invalid input
        '401':
          description: Unauthorized

  /accounts/{accountId}/events/{eventId}/participants/upload:
    post:
      summary: Upload participants for an event
      operationId: uploadParticipants
      tags:
        - participants
      parameters:
        - name: accountId
          in: path
          required: true
          schema:
            type: integer
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Participants uploaded successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  imported:
                    type: integer
        '400':
          description: Invalid input
        '401':
          description: Unauthorized

  /accounts/{accountId}/events/{eventId}/sendInvitations:
    post:
      summary: Send invitations to event participants
      operationId: sendInvitations
      tags:
        - invitations
      parameters:
        - name: accountId
          in: path
          required: true
          schema:
            type: integer
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Invitations sent successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
        '401':
          description: Unauthorized
        '404':
          description: Event not found

components:
  schemas:
    EventCreation:
      type: object
      required:
        - name
        - database
      properties:
        name:
          type: string
        database:
          type: string
          enum: [GoogleSpreadsheet, OwnDB]

    Event:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        database:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    Participant:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        qrCode:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

security:
  - bearerAuth: []